✨ plugin api changes
