+++ lunchable/_cli.py
@@ -2,14 +2,15 @@
 Lunchmoney CLI
 """
 
-import datetime
 import logging
 import sys
 from json import JSONDecodeError
-from typing import Any, Dict, Optional, Union
+from typing import Any, Dict, Optional
 
 import click
 import httpx
+from click_plugins import with_plugins
+from importlib_metadata import entry_points
 from pydantic_core import to_jsonable_python
 from rich import print, print_json, traceback
 
@@ -70,6 +71,11 @@ def transactions() -> None:
 def plugins() -> None:
     """
     Interact with Lunchable Plugins
+
+    Install lunchable with the "plugins" extra to get
+    all the known plugins
+
+    pipx install "lunchable[plugins]"
     """
 
 
@@ -147,246 +153,6 @@ def lunchmoney_transactions(
     print_json(data=json_data)
 
 
-@plugins.group()
-def splitlunch() -> None:
-    """
-    Splitwise Plugin for lunchable, SplitLunch 💲🍱
-    """
-    pass
-
-
-dated_after = click.option(
-    "--dated-after",
-    default=None,
-    help="ISO 8601 Date time. Return expenses later that this date",
-)
-dated_before = click.option(
-    "--dated-before",
-    default=None,
-    help="ISO 8601 Date time. Return expenses earlier than this date",
-)
-
-
-@splitlunch.command("expenses")
-@click.option(
-    "--limit", default=None, help="Limit the amount of Results. 0 returns everything."
-)
-@click.option("--offset", default=None, help="Number of expenses to be skipped")
-@click.option("--limit", default=None, help="Number of expenses to be returned")
-@click.option("--group-id", default=None, help="GroupID of the expenses")
-@click.option("--friendship-id", default=None, help="FriendshipID of the expenses")
-@dated_after
-@dated_before
-@click.option(
-    "--updated-after",
-    default=None,
-    help="ISO 8601 Date time. Return expenses updated after this date",
-)
-@click.option(
-    "--updated-before",
-    default=None,
-    help="ISO 8601 Date time. Return expenses updated before this date",
-)
-def splitlunch_expenses(**kwargs: Union[int, str, bool]) -> None:
-    """
-    Retrieve Splitwise Expenses
-    """
-    from lunchable.plugins.splitlunch import SplitLunch
-
-    splitlunch = SplitLunch()
-    if set(kwargs.values()) == {None}:
-        kwargs["limit"] = 5
-    expenses = splitlunch.get_expenses(**kwargs)  # type: ignore[arg-type]
-    json_data = to_jsonable_python(expenses)
-    print_json(data=json_data)
-
-
-tag_transactions = click.option(
-    "--tag-transactions",
-    is_flag=True,
-    help="Tag the resulting transactions with a `Splitwise` tag.",
-)
-financial_partner_id = click.option(
-    "--financial-partner-id",
-    default=None,
-    type=click.INT,
-    help="Splitwise ID of your financial partner.",
-)
-financial_partner_email = click.option(
-    "--financial-partner-email",
-    default=None,
-    help="Splitwise Email Address of your financial partner.",
-)
-financial_partner_group_id = click.option(
-    "--financial-partner-group-id",
-    default=None,
-    type=click.INT,
-    help="Splitwise Group ID for financial partner transactions.",
-)
-
-
-@splitlunch.command("splitlunch")
-@tag_transactions
-def make_splitlunch(**kwargs: Union[int, str, bool]) -> None:
-    """
-    Split all `SplitLunch` tagged transactions in half.
-
-    One of these new splits will be recategorized to `Reimbursement`.
-    """
-    from lunchable.plugins.splitlunch import SplitLunch
-
-    splitlunch = SplitLunch()
-    results = splitlunch.make_splitlunch(**kwargs)  # type: ignore[arg-type]
-    json_data = to_jsonable_python(results)
-    print_json(data=json_data)
-
-
-@splitlunch.command("splitlunch-import")
-@tag_transactions
-@financial_partner_id
-@financial_partner_email
-@financial_partner_group_id
-def make_splitlunch_import(**kwargs: Union[int, str, bool]) -> None:
-    """
-    Import `SplitLunchImport` tagged transactions to Splitwise and Split them in Lunch Money
-
-    Send a transaction to Splitwise and then split the original transaction in Lunch Money.
-    One of these new splits will be recategorized to `Reimbursement`. Any tags will be
-    reapplied.
-    """
-    from lunchable.plugins.splitlunch import SplitLunch
-
-    financial_partner_id: Optional[int] = kwargs.pop("financial_partner_id")  # type: ignore[assignment]
-    financial_partner_email: Optional[str] = kwargs.pop("financial_partner_email")  # type: ignore[assignment]
-    financial_partner_group_id: Optional[int] = kwargs.pop("financial_partner_group_id")  # type: ignore[assignment]
-    splitlunch = SplitLunch(
-        financial_partner_id=financial_partner_id,
-        financial_partner_email=financial_partner_email,
-        financial_partner_group_id=financial_partner_group_id,
-    )
-    results = splitlunch.make_splitlunch_import(**kwargs)  # type: ignore[arg-type]
-    json_data = to_jsonable_python(results)
-    print_json(data=json_data)
-
-
-@splitlunch.command("splitlunch-direct-import")
-@tag_transactions
-@financial_partner_id
-@financial_partner_email
-@financial_partner_group_id
-def make_splitlunch_direct_import(**kwargs: Union[int, str, bool]) -> None:
-    """
-    Import `SplitLunchDirectImport` tagged transactions to Splitwise and Split them in Lunch Money
-
-    Send a transaction to Splitwise and then split the original transaction in Lunch Money.
-    One of these new splits will be recategorized to `Reimbursement`. Any tags will be
-    reapplied.
-    """
-    from lunchable.plugins.splitlunch import SplitLunch
-
-    financial_partner_id: Optional[int] = kwargs.pop("financial_partner_id")  # type: ignore[assignment]
-    financial_partner_email: Optional[str] = kwargs.pop("financial_partner_email")  # type: ignore[assignment]
-    financial_partner_group_id: Optional[int] = kwargs.pop("financial_partner_group_id")  # type: ignore[assignment]
-    splitlunch = SplitLunch(
-        financial_partner_id=financial_partner_id,
-        financial_partner_email=financial_partner_email,
-        financial_partner_group_id=financial_partner_group_id,
-    )
-    results = splitlunch.make_splitlunch_direct_import(**kwargs)  # type: ignore[arg-type]
-    json_data = to_jsonable_python(results)
-    print_json(data=json_data)
-
-
-@splitlunch.command("update-balance")
-def update_splitwise_balance() -> None:
-    """
-    Update the Splitwise Asset Balance
-    """
-    from lunchable.plugins.splitlunch import SplitLunch
-
-    splitlunch = SplitLunch()
-    updated_asset = splitlunch.update_splitwise_balance()
-    json_data = to_jsonable_python(updated_asset)
-    print_json(data=json_data)
-
-
-@splitlunch.command("refresh")
-@dated_after
-@dated_before
-@click.option(
-    "--allow-self-paid/--no-allow-self-paid",
-    default=False,
-    help="Allow self-paid expenses to be imported (filtered out by default).",
-)
-@click.option(
-    "--allow-payments/--no-allow-payments",
-    default=False,
-    help="Allow payments to be imported (filtered out by default).",
-)
-def refresh_splitwise_transactions(
-    dated_before: Optional[datetime.datetime],
-    dated_after: Optional[datetime.datetime],
-    allow_self_paid: bool,
-    allow_payments: bool,
-) -> None:
-    """
-    Import New Splitwise Transactions to Lunch Money and
-
-    This function gets all transactions from Splitwise, all transactions from
-    your Lunch Money Splitwise account and compares the two. This also updates
-    the account balance.
-    """
-    from lunchable.plugins.splitlunch import SplitLunch
-
-    splitlunch = SplitLunch()
-    response = splitlunch.refresh_splitwise_transactions(
-        dated_before=dated_before,
-        dated_after=dated_after,
-        allow_self_paid=allow_self_paid,
-        allow_payments=allow_payments,
-    )
-    json_data = to_jsonable_python(response)
-    print_json(data=json_data)
-
-
-@plugins.group()
-def pushlunch() -> None:
-    """
-    Push Notifications for Lunch Money: PushLunch 📲
-    """
-    pass
-
-
-@pushlunch.command("notify")
-@click.option(
-    "--continuous",
-    is_flag=True,
-    help="Whether to continuously check for more uncleared transactions, "
-    "waiting a fixed amount in between checks.",
-)
-@click.option(
-    "--interval",
-    default=None,
-    help="Sleep Interval in Between Tries - only applies if `continuous` is set. "
-    "Defaults to 60 (minutes). Cannot be less than 5 (minutes)",
-)
-@click.option(
-    "--user-key",
-    default=None,
-    help="Pushover User Key. Defaults to `PUSHOVER_USER_KEY` env var",
-)
-def notify(continuous: bool, interval: int, user_key: str) -> None:
-    """
-    Send a Notification for each Uncleared Transaction
-    """
-    from lunchable.plugins.pushlunch import PushLunch
-
-    push = PushLunch(user_key=user_key)
-    if interval is not None:
-        interval = int(interval)
-    push.notify_uncleared_transactions(continuous=continuous, interval=interval)
-
-
 @cli.command()
 @click.argument("URL")
 @click.option("-X", "--request", default="GET", help="Specify request command to use")
@@ -423,58 +189,5 @@ def http(context: LunchMoneyContext, url: str, request: str, data: str) -> None:
     print_json(data=json_data)
 
 
-@plugins.group()
-def primelunch() -> None:
-    """
-    PrimeLunch CLI - Syncing LunchMoney with Amazon
-    """
-
-
-@primelunch.command("run")
-@click.option(
-    "-f",
-    "--file",
-    "csv_file",
-    type=click.Path(exists=True, resolve_path=True),
-    help="File Path of the Amazon Export",
-    required=True,
-)
-@click.option(
-    "-w",
-    "--window",
-    "window",
-    type=click.INT,
-    help="Allowable time window between Amazon transaction date and "
-    "credit card transaction date",
-    default=7,
-)
-@click.option(
-    "-a",
-    "--all",
-    "update_all",
-    is_flag=True,
-    type=click.BOOL,
-    help="Whether to skip the confirmation step and simply update all matched "
-    "transactions",
-    default=False,
-)
-@click.option(
-    "-t",
-    "--token",
-    "access_token",
-    type=click.STRING,
-    help="LunchMoney Access Token - defaults to the LUNCHMONEY_ACCESS_TOKEN environment variable",
-    envvar="LUNCHMONEY_ACCESS_TOKEN",
-)
-def run_primelunch(
-    csv_file: str, window: int, update_all: bool, access_token: str
-) -> None:
-    """
-    Run the PrimeLunch Update Process
-    """
-    from lunchable.plugins.primelunch.primelunch import PrimeLunch
-
-    primelunch = PrimeLunch(
-        file_path=csv_file, time_window=window, access_token=access_token
-    )
-    primelunch.process_transactions(confirm=not update_all)
+discovered_plugins = entry_points(group="lunchable.cli")
+with_plugins(discovered_plugins)(plugins)
