--- lunchable/plugins/base/__init__.py
@@ -1,11 +1,0 @@
-"""
-Base Apps for Plugins
-"""
-
-from .base_app import LunchableApp, LunchableModelType, LunchableTransactionsApp
-
-__all__ = [
-    "LunchableTransactionsApp",
-    "LunchableApp",
-    "LunchableModelType",
-]
