+++ lunchable/plugins/pushlunch/__init__.py
@@ -1,9 +1,3 @@
 """
 PushLunch lunchable Plugin
 """
-
-from .pushover import PushLunch
-
-__all__ = [
-    "PushLunch",
-]
