+++ pyproject.toml
@@ -23,7 +23,9 @@ dependencies = [
   "click>=8.0.1",
   "pydantic>=2,<3",
   "rich>=10.0.0",
-  "httpx"
+  "httpx",
+  "importlib_metadata>=3.6",
+  "click-plugins>=1.1.1"
 ]
 description = "A simple Python SDK around the Lunch Money Developer API"
 dynamic = ["version"]
@@ -33,14 +35,21 @@ name = "lunchable"
 readme = "README.md"
 requires-python = ">=3.8,<4"
 
+[project.entry-points."lunchable.cli"]
+primelunch = "lunchable.plugins.primelunch.cli:primelunch"
+pushlunch = "lunchable.plugins.pushlunch.cli:pushlunch"
+splitlunch = "lunchable.plugins.splitlunch.cli:splitlunch"
+
 [project.optional-dependencies]
 all = [
   "pandas",
   "python-dateutil",
   "splitwise>=2.3.0,<3.0.0"
 ]
-apps = [
-  "pandas"
+plugins = [
+  "pandas",
+  "python-dateutil",
+  "splitwise>=2.3.0,<3.0.0"
 ]
 primelunch = [
   "pandas"
